Wire templates into the active /lists pages (not legacy /shopping-lists)
The navigation points at /lists, so the create dialog, "Save as template"
action, and "From template" indicator belong there. Moves the template UI
to /lists and /lists/[id] and restores the unused legacy /shopping-lists
pages to their original state.
