+++ anything-frontend/src/app/shopping-lists/page.test.tsx
@@ -7,8 +7,6 @@ import { toast } from 'sonner'
 // Mock the apiClient module
 const mockShoppingListsGet = jest.fn()
 const mockShoppingListsPost = jest.fn()
-const mockTemplatesGet = jest.fn()
-const mockFromTemplatePost = jest.fn()
 
 jest.mock('@/lib/apiClient', () => ({
   apiClient: {
@@ -16,12 +14,6 @@ jest.mock('@/lib/apiClient', () => ({
       checklists: {
         get: (...args: unknown[]) => mockShoppingListsGet(...args),
         post: (...args: unknown[]) => mockShoppingListsPost(...args),
-        templates: {
-          get: (...args: unknown[]) => mockTemplatesGet(...args),
-        },
-        fromTemplate: {
-          post: (...args: unknown[]) => mockFromTemplatePost(...args),
-        },
       },
     },
   },
@@ -51,7 +43,6 @@ describe('ShoppingListsPage', () => {
     jest.clearAllMocks()
     localStorage.setItem('user', JSON.stringify({ email: 'test@test.com', name: 'Test User', role: 'User' }))
     localStorage.setItem('accessToken', 'test-token')
-    mockTemplatesGet.mockResolvedValue([])
   })
 
   afterEach(() => {
@@ -101,7 +92,7 @@ describe('ShoppingListsPage', () => {
     })
   })
 
-  it('should open create dialog when plus button is clicked', async () => {
+  it('should open create form when plus button is clicked', async () => {
     const user = userEvent.setup()
     mockShoppingListsGet.mockResolvedValue([])
 
@@ -110,13 +101,12 @@ describe('ShoppingListsPage', () => {
     const plusButton = screen.getByRole('button', { name: 'Create shopping list' })
     await user.click(plusButton)
 
-    expect(screen.getByRole('dialog')).toBeInTheDocument()
     expect(screen.getByPlaceholderText('List name...')).toBeInTheDocument()
   })
 
-  it('should create a checklist (type 0) by default and navigate to it', async () => {
+  it('should create a new list and navigate to it', async () => {
     const user = userEvent.setup()
-    const mockNewList = { id: 3, name: 'Chores', createdOn: '2024-01-03T00:00:00Z' }
+    const mockNewList = { id: 3, name: 'Party Supplies', createdOn: '2024-01-03T00:00:00Z' }
     mockShoppingListsGet.mockResolvedValue([])
     mockShoppingListsPost.mockResolvedValueOnce(mockNewList)
 
@@ -124,65 +114,20 @@ describe('ShoppingListsPage', () => {
 
     await user.click(screen.getByRole('button', { name: 'Create shopping list' }))
 
-    await user.type(screen.getByPlaceholderText('List name...'), 'Chores')
-    await user.click(screen.getByRole('button', { name: 'Create' }))
+    const input = screen.getByPlaceholderText('List name...')
+    await user.type(input, 'Party Supplies')
 
-    await waitFor(() => {
-      expect(mockShoppingListsPost).toHaveBeenCalledWith({ name: 'Chores', type: 0 })
-    })
-
-    expect(toast.success).toHaveBeenCalledWith('Checklist created')
-    expect(mockPush).toHaveBeenCalledWith('/shopping-lists/3')
-  })
-
-  it('should create a shopping list (type 1) when that mode is selected', async () => {
-    const user = userEvent.setup()
-    const mockNewList = { id: 4, name: 'Party Supplies', createdOn: '2024-01-03T00:00:00Z' }
-    mockShoppingListsGet.mockResolvedValue([])
-    mockShoppingListsPost.mockResolvedValueOnce(mockNewList)
-
-    render(<ShoppingListsPage />)
-
-    await user.click(screen.getByRole('button', { name: 'Create shopping list' }))
-    await user.click(screen.getByRole('button', { name: 'Shopping list' }))
-    await user.type(screen.getByPlaceholderText('List name...'), 'Party Supplies')
-    await user.click(screen.getByRole('button', { name: 'Create' }))
+    await user.click(screen.getByRole('button', { name: 'Create list' }))
 
     await waitFor(() => {
       expect(mockShoppingListsPost).toHaveBeenCalledWith({ name: 'Party Supplies', type: 1 })
     })
 
     expect(toast.success).toHaveBeenCalledWith('Shopping list created')
-    expect(mockPush).toHaveBeenCalledWith('/shopping-lists/4')
+    expect(mockPush).toHaveBeenCalledWith('/shopping-lists/3')
   })
 
-  it('should create a list from a template', async () => {
-    const user = userEvent.setup()
-    mockShoppingListsGet.mockResolvedValue([])
-    mockTemplatesGet.mockResolvedValue([{ id: 7, name: 'Weekly Groceries', type: 1, itemCount: 5 }])
-    mockFromTemplatePost.mockResolvedValueOnce({ id: 9, name: 'Weekly Groceries', createdOn: '2024-01-03T00:00:00Z' })
-
-    render(<ShoppingListsPage />)
-
-    await user.click(screen.getByRole('button', { name: 'Create shopping list' }))
-    await user.click(screen.getByRole('button', { name: 'From template' }))
-
-    await waitFor(() => {
-      expect(screen.getByText('Weekly Groceries')).toBeInTheDocument()
-    })
-
-    await user.click(screen.getByRole('button', { name: /Weekly Groceries/ }))
-    await user.click(screen.getByRole('button', { name: 'Create' }))
-
-    await waitFor(() => {
-      expect(mockFromTemplatePost).toHaveBeenCalledWith({ templateId: 7, name: 'Weekly Groceries' })
-    })
-
-    expect(toast.success).toHaveBeenCalledWith('List created from template')
-    expect(mockPush).toHaveBeenCalledWith('/shopping-lists/9')
-  })
-
-  it('should disable Create when list name is empty', async () => {
+  it('should not submit when list name is empty', async () => {
     const user = userEvent.setup()
     mockShoppingListsGet.mockResolvedValue([])
 
@@ -190,7 +135,8 @@ describe('ShoppingListsPage', () => {
 
     await user.click(screen.getByRole('button', { name: 'Create shopping list' }))
 
-    expect(screen.getByRole('button', { name: 'Create' })).toBeDisabled()
+    await user.click(screen.getByRole('button', { name: 'Create list' }))
+
     expect(mockShoppingListsPost).not.toHaveBeenCalled()
   })
 
@@ -205,14 +151,14 @@ describe('ShoppingListsPage', () => {
 
     await user.type(screen.getByPlaceholderText('List name...'), 'Fail List')
 
-    await user.click(screen.getByRole('button', { name: 'Create' }))
+    await user.click(screen.getByRole('button', { name: 'Create list' }))
 
     await waitFor(() => {
-      expect(toast.error).toHaveBeenCalledWith('Failed to create list. Please try again.')
+      expect(toast.error).toHaveBeenCalledWith('Failed to create shopping list. Please try again.')
     })
   })
 
-  it('should close the dialog when Cancel is clicked', async () => {
+  it('should cancel create form when Cancel is clicked', async () => {
     const user = userEvent.setup()
     mockShoppingListsGet.mockResolvedValue([])
 
@@ -223,9 +169,7 @@ describe('ShoppingListsPage', () => {
 
     await user.click(screen.getByRole('button', { name: 'Cancel' }))
 
-    await waitFor(() => {
-      expect(screen.queryByPlaceholderText('List name...')).not.toBeInTheDocument()
-    })
+    expect(screen.queryByPlaceholderText('List name...')).not.toBeInTheDocument()
   })
 
   it('should navigate to list when row is clicked', async () => {
@@ -256,7 +200,7 @@ describe('ShoppingListsPage', () => {
 
     await user.click(screen.getByRole('button', { name: 'Create shopping list' }))
     await user.type(screen.getByPlaceholderText('List name...'), 'Test')
-    await user.click(screen.getByRole('button', { name: 'Create' }))
+    await user.click(screen.getByRole('button', { name: 'Create list' }))
 
     expect(screen.getByRole('button', { name: 'Creating...' })).toBeInTheDocument()
   })
