+++ anything-frontend/src/app/shopping-lists/page.tsx
@@ -2,9 +2,9 @@
 
 import { Button } from "@/components/ui/button";
 import { CountBadge } from "@/components/ui/count-badge";
-import { useShoppingLists, useReorderShoppingLists } from "@/hooks/useShoppingLists";
-import { CreateListDialog } from "@/components/CreateListDialog";
-import { useState, useEffect } from "react";
+import { useShoppingLists, useCreateShoppingList, useReorderShoppingLists } from "@/hooks/useShoppingLists";
+import { useState, useEffect, useRef } from "react";
+import { toast } from "sonner";
 import { useRouter } from "next/navigation";
 import { useHeaderActions } from "@/context/PageActionsContext";
 import { PageTitle } from "@/components/PageTitle";
@@ -71,7 +71,10 @@ function DraggableShoppingListItem({ list, onClick }: { list: ShoppingListRespon
 
 export default function ShoppingListsPage() {
   const [isCreating, setIsCreating] = useState(false);
+  const [newListName, setNewListName] = useState("");
+  const inputRef = useRef<HTMLInputElement>(null);
   const { data: lists, isLoading, error } = useShoppingLists();
+  const createList = useCreateShoppingList();
   const reorderLists = useReorderShoppingLists();
   const router = useRouter();
   const { setHeaderActions } = useHeaderActions();
@@ -95,6 +98,27 @@ export default function ShoppingListsPage() {
     reorderLists.mutate(reordered.map((l) => l.id ?? 0));
   };
 
+  const handleCreateList = async (e: React.FormEvent) => {
+    e.preventDefault();
+    if (!newListName.trim()) return;
+    try {
+      const newList = await createList.mutateAsync({ name: newListName });
+      setNewListName("");
+      setIsCreating(false);
+      toast.success("Shopping list created");
+      if (newList?.id) {
+        router.push(`/shopping-lists/${newList.id}`);
+      }
+    } catch {
+      toast.error("Failed to create shopping list. Please try again.");
+    }
+  };
+
+  const handleCancelCreate = () => {
+    setIsCreating(false);
+    setNewListName("");
+  };
+
   useEffect(() => {
     setHeaderActions(
       <div className="flex items-center gap-1 ml-auto">
@@ -111,16 +135,40 @@ export default function ShoppingListsPage() {
     return () => setHeaderActions(null);
   }, [setHeaderActions]);
 
+  useEffect(() => {
+    if (isCreating) {
+      inputRef.current?.focus();
+    }
+  }, [isCreating]);
+
   const hasActiveLists = lists && lists.length > 0;
 
   return (
     <div className="container mx-auto px-4 py-4 max-w-4xl">
       <PageTitle>Shopping Lists</PageTitle>
-      <CreateListDialog
-        open={isCreating}
-        onOpenChange={setIsCreating}
-        onCreated={(listId) => router.push(`/shopping-lists/${listId}`)}
-      />
+      {isCreating && (
+        <form onSubmit={handleCreateList} className="mb-4">
+          <div className="flex gap-2">
+            <input
+              ref={inputRef}
+              type="text"
+              value={newListName}
+              onChange={(e) => setNewListName(e.target.value)}
+              placeholder="List name..."
+              className="flex-1 px-3 py-2 text-sm border border-gray-300 dark:border-gray-600 rounded-lg focus:outline-none focus:ring-2 focus:ring-blue-500 dark:bg-gray-700 dark:text-white"
+              onKeyDown={(e) => {
+                if (e.key === "Escape") handleCancelCreate();
+              }}
+            />
+            <Button type="submit" size="sm" disabled={createList.isPending}>
+              {createList.isPending ? "Creating..." : "Create list"}
+            </Button>
+            <Button type="button" size="sm" variant="ghost" onClick={handleCancelCreate}>
+              Cancel
+            </Button>
+          </div>
+        </form>
+      )}
 
       {isLoading && (
         <div className="text-center py-12 text-gray-500 dark:text-gray-400">
@@ -134,7 +182,7 @@ export default function ShoppingListsPage() {
         </div>
       )}
 
-      {lists?.length === 0 && (
+      {lists && lists.length === 0 && !isCreating && (
         <div className="text-center py-12 text-gray-500 dark:text-gray-400">
           No shopping lists yet.
         </div>
